Final refactoring
