+++ lib/vite-public-config.js
@@ -1,0 +1,13 @@
+module.exports = async ({
+  input
+}) => {
+
+  /** @type {import('vite').UserConfig} */
+  return {
+    build: {
+      rollupOptions: {
+        input
+      }
+    }
+  };
+};
