+++ vite.mjs
@@ -1,0 +1,2 @@
+export * from 'vite';
+export * as default from 'vite';
